rename index.ts -> main.ts
