+++ practice-typescript/src/index.html
@@ -91,6 +91,6 @@
       <button type="submit" class="btn btn-update">Create User</button>
     </form>
     <div class="toast" id="toast">This is a simple toast message.</div>
-    <script type="module" src="./scripts/index.ts"></script>
+    <script type="module" src="./scripts/main.ts"></script>
   </body>
 </html>
