--- practice-typescript/src/scripts/constants/HTTPMethod.ts
@@ -1,8 +1,0 @@
-enum HTTPMethod {
-  GET = 'GET',
-  POST = 'POST',
-  PUT = 'PUT',
-  DELETE = 'DELETE',
-}
-
-export default HTTPMethod;
