--- practice-typescript/src/scripts/constants/action.ts
@@ -1,6 +1,0 @@
-enum ACTION {
-  CREATE = 'Create User',
-  UPDATE = 'Update User',
-}
-
-export { ACTION };
