--- practice-typescript/src/scripts/constants/api.ts
@@ -1,4 +1,0 @@
-const BASE_URL = 'https://6645cb84b8925626f89336b4.mockapi.io';
-const USERS = 'users';
-
-export { BASE_URL, USERS };
