--- practice-typescript/src/scripts/constants/message.ts
@@ -1,20 +1,0 @@
-enum MESSAGE_ERROR {
-  EMAIL = 'Please enter a valid email address',
-  FIRST_NAME = 'Please enter your first name',
-  LAST_NAME = 'Please enter your last name',
-  SPECIAL_CHARACTER = 'should not include a special character',
-  GREATER_THAN_LENGTH = 'cannot exceed 12 characters',
-  PHONE_NUMBER = 'Please enter your phone format 0-123-456-789',
-  EMPTY = '',
-  ACCOUNT_EXIST = 'Username already exits',
-}
-
-enum MESSAGE_SUCCESS {
-  CREATE_SUCCESS = 'Create success a new user',
-  UPDATE_SUCCESS = 'Update success a user',
-  DELETE_SUCCESS = 'Delete success user',
-}
-
-const TIME_MESSAGE = 3 * 1000;
-
-export { MESSAGE_ERROR, MESSAGE_SUCCESS, TIME_MESSAGE };
