--- practice-typescript/src/scripts/constants/notification.ts
@@ -1,5 +1,0 @@
-const NO_USERS = 'There is no data to display';
-
-const USER_NOT_FOUND = 'User was not found according to search data';
-
-export { NO_USERS, USER_NOT_FOUND };
