+++ practice-typescript/src/scripts/helpers/toast.ts
@@ -1,11 +1,11 @@
-import { TIME_MESSAGE } from 'scripts/constants/message';
+import { TIME_MESSAGE, SHOW_TOAST } from 'scripts/constants/constants';
 
 const toastMessage = (toast: HTMLElement, message: string, status: string) => {
-  toast.classList.add('toast__show');
+  toast.classList.add(SHOW_TOAST);
   toast.classList.add(status);
   toast.textContent = message;
   setTimeout(() => {
-    toast.classList.remove('toast__show');
+    toast.classList.remove(SHOW_TOAST);
   }, TIME_MESSAGE);
   setTimeout(() => {
     toast.classList.remove(status);
