+++ practice-typescript/src/scripts/helpers/validate.ts
@@ -1,5 +1,5 @@
-import { FIRST_NAME, LAST_NAME, MAX_LENGTH } from 'scripts/constants/constant';
-import { MESSAGE_ERROR } from 'scripts/constants/message';
+import { FIRST_NAME, LAST_NAME, MAX_LENGTH } from 'scripts/constants/constants';
+import { MESSAGE_ERROR } from 'scripts/constants/constants';
 import { createElement } from 'scripts/helpers/createElement';
 
 const validationFunctions = {
@@ -17,11 +17,11 @@ const validationFunctions = {
   },
   emailError: (email: string) =>
     !validationFunctions.validateEmail(email)
-      ? MESSAGE_ERROR.EMAIL
+      ? MESSAGE_ERROR.EMAIL_EMPTY
       : MESSAGE_ERROR.EMPTY,
   nameError(name: string) {
     if (!name) {
-      return MESSAGE_ERROR.FIRST_NAME;
+      return MESSAGE_ERROR.EMAIL_EMPTY;
     }
     if (name.length > MAX_LENGTH) {
       return FIRST_NAME + MESSAGE_ERROR.GREATER_THAN_LENGTH;
@@ -33,7 +33,7 @@ const validationFunctions = {
   },
   lastNameError(name: string) {
     if (!name) {
-      return MESSAGE_ERROR.LAST_NAME;
+      return MESSAGE_ERROR.LAST_NAME_EMPTY;
     }
     if (name.length > MAX_LENGTH) {
       return LAST_NAME + MESSAGE_ERROR.GREATER_THAN_LENGTH;
