+++ practice-typescript/src/scripts/main.ts
